+++ CHANGES.md
@@ -1,9 +1,9 @@
 # List of changes to be made in the 2.0 release
-- change drawer search to form, throw redirect on server, consider enhancing
+- remove routeID, add wheelchair and full to supported fields
+ - either use a new vehicle api or adjust current one to GTT standard (72B, ...)
 - change strike to server load, load everything in parent load, reuse in component, remove API
 - consider moving all pollers to separate library directory
 - investigate possibility of using enhanced forms with method="GET" with search queries
 
 # Furter changes
-- move pollers to graphql
 - move to a different map library